profile: move MatProfile tests into separate test file so they only run when custom profiler active
